# apps/api-gateway/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Servicienta API Gateway
  version: 0.1.0
  description: |
    OpenAPI specification for the current Express API Gateway.

    Notes:
    - Most application responses use the `{ data: ... }` envelope.
    - Error responses use `{ error, requestId }`.
    - `/health` is the only endpoint that does not use the standard envelope.
servers:
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/orders/{orderId}/operations:
    post:
      tags: [Operations]
      summary: Create operation for an order
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperationInput'
            examples:
              createOperation:
                value:
                  technician_id: 0f095605-b76a-45c5-8bd6-0551d3845d22
                  scheduled_at: '2026-05-04T15:00:00.000Z'
      responses:
        '201':
          $ref: '#/components/responses/OperationResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidState:
                  value:
                    error: Order is not accepted for a new operation
                    requestId: req_123
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                clientRoleRequired:
                  value:
                    error: Client role required
                    requestId: req_123
        '404':
          description: Related entity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                orderNotFound:
                  value:
                    error: Order not found
                    requestId: req_123
                technicianNotFound:
                  value:
                    error: Technician profile not found
                    requestId: req_123
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/operations/current:
    get:
      tags: [Operations]
      summary: List current operations for authenticated client or technician
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OperationStatus'
      responses:
        '200':
          description: Paginated current operations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedOperationsResponse'
              examples:
                success:
                  value:
                    data:
                      items:
                        - id: d364272c-aa8c-4273-86b0-6fd9999bb222
                          order_id: eaf86abc-ae80-486a-a0cf-6cacbe761111
                          technician_id: 0f095605-b76a-45c5-8bd6-0551d3845d22
                          status: scheduled
                          scheduled_at: '2026-05-04T15:00:00.000Z'
                          description: Cambio de termostato coordinado
                          technician_completed_at: null
                          completed_at: null
                          created_at: '2026-05-03T14:05:00.000Z'
                          updated_at: '2026-05-03T14:05:00.000Z'
                      pagination:
                        page: 1
                        pageSize: 25
                        total: 1
                        totalPages: 1
                      summary:
                        totalOperations: 1
                        pendingOperations: 0
                        scheduledOperations: 1
                        completedTechOperations: 0
                        completedOperations: 0
                        cancelledOperations: 0
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                roleRequired:
                  value:
                    error: Client or technician role required
                    requestId: req_123
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/operations/{operationId}:
    get:
      tags: [Operations]
      summary: Get current operation by id for authenticated client or technician
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        '200':
          $ref: '#/components/responses/OperationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                roleRequired:
                  value:
                    error: Client or technician role required
                    requestId: req_123
        '404':
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                operationNotFound:
                  value:
                    error: Operation not found
                    requestId: req_123
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/operations/{operationId}/complete:
    post:
      tags: [Operations]
      summary: Complete operation as technician (legacy alias for complete-tech)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        '200':
          $ref: '#/components/responses/OperationResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                alreadyCompleted:
                  value:
                    error: Operation is already completed
                    requestId: req_123
                cancelledOperation:
                  value:
                    error: Cancelled operations cannot be completed
                    requestId: req_123
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                technicianRoleRequired:
                  value:
                    error: Technician role required
                    requestId: req_123
        '404':
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                operationNotFound:
                  value:
                    error: Operation not found
                    requestId: req_123
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/operations/{operationId}/schedule:
    post:
      tags: [Operations]
      summary: Schedule pending operation as technician
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OperationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleOperationInput'
      responses:
        '200':
          $ref: '#/components/responses/OperationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/operations/{operationId}/complete-tech:
    post:
      tags: [Operations]
      summary: Mark scheduled operation completed by technician
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        '200':
          $ref: '#/components/responses/OperationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/operations/{operationId}/confirm-completed:
    post:
      tags: [Operations]
      summary: Confirm technician-completed operation as client
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        '200':
          $ref: '#/components/responses/OperationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/operations/{operationId}/cancel:
    post:
      tags: [Operations]
      summary: Cancel pending or scheduled operation as technician
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        '200':
          $ref: '#/components/responses/OperationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/operations:
    get:
      tags: [Operations]
      summary: List operations for admin
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OperationStatus'
      responses:
        '200':
          description: Paginated admin operations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAdminOperationsResponse'
              examples:
                success:
                  value:
                    data:
                      items:
                        - id: d364272c-aa8c-4273-86b0-6fd9999bb222
                          order_id: eaf86abc-ae80-486a-a0cf-6cacbe761111
                          technician_id: 0f095605-b76a-45c5-8bd6-0551d3845d22
                          status: completed
                          scheduled_at: '2026-05-04T15:00:00.000Z'
                          description: Cambio de termostato
                          technician_completed_at: '2026-05-04T16:10:00.000Z'
                          completed_at: '2026-05-04T16:30:00.000Z'
                          created_at: '2026-05-03T14:05:00.000Z'
                          updated_at: '2026-05-04T16:30:00.000Z'
                          order_status: completed
                          order_flow_type: client_selects
                          service_address_text: Calle Falsa 123
                          client_id: 1f0fe2c1-e1ff-4f4a-a965-14bc3cb99631
                          client_email: client@example.com
                          client_name: Kiara
                          client_surname: Pérez
                          technician_email: technician@example.com
                          technician_name: Marcelo
                          technician_surname: Ruiz
                          technician_public_slug: tech-abc12345
                      pagination:
                        page: 1
                        pageSize: 25
                        total: 1
                        totalPages: 1
                      summary:
                        totalOperations: 1
                        pendingOperations: 0
                        scheduledOperations: 0
                        completedTechOperations: 0
                        completedOperations: 1
                        cancelledOperations: 0
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/admin/operations/{operationId}:
    get:
      tags: [Operations]
      summary: Get operation by id for admin
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OperationId'
      responses:
        '200':
          $ref: '#/components/responses/AdminOperationResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Operation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                operationNotFound:
                  value:
                    error: Operation not found
                    requestId: req_123
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        $ref: '#/components/schemas/UsersPageSize'
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OperationId:
      name: operationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    OperationResponse:
      description: Operation response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OperationResponse'
          examples:
            success:
              value:
                data:
                  id: d364272c-aa8c-4273-86b0-6fd9999bb222
                  order_id: eaf86abc-ae80-486a-a0cf-6cacbe761111
                  technician_id: 0f095605-b76a-45c5-8bd6-0551d3845d22
                  status: scheduled
                  scheduled_at: '2026-05-04T15:00:00.000Z'
                  description: Cambio de termostato coordinado
                  technician_completed_at: null
                  completed_at: null
                  created_at: '2026-05-03T14:05:00.000Z'
                  updated_at: '2026-05-03T14:05:00.000Z'
    AdminOperationResponse:
      description: Admin operation response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminOperationResponse'
          examples:
            success:
              value:
                data:
                  id: d364272c-aa8c-4273-86b0-6fd9999bb222
                  order_id: eaf86abc-ae80-486a-a0cf-6cacbe761111
                  technician_id: 0f095605-b76a-45c5-8bd6-0551d3845d22
                  status: completed
                  scheduled_at: '2026-05-04T15:00:00.000Z'
                  description: Cambio de termostato
                  technician_completed_at: '2026-05-04T16:10:00.000Z'
                  completed_at: '2026-05-04T16:30:00.000Z'
                  created_at: '2026-05-03T14:05:00.000Z'
                  updated_at: '2026-05-04T16:30:00.000Z'
                  order_status: completed
                  order_flow_type: client_selects
                  service_address_text: Calle Falsa 123
                  client_id: 1f0fe2c1-e1ff-4f4a-a965-14bc3cb99631
                  client_email: client@example.com
                  client_name: Kiara
                  client_surname: Pérez
                  technician_email: technician@example.com
                  technician_name: Marcelo
                  technician_surname: Ruiz
                  technician_public_slug: tech-abc12345
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidPage:
              value:
                error: Invalid page
                requestId: req_123
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                error: Missing bearer token
                requestId: req_123
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                error: Admin role required
                requestId: req_123
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internal:
              value:
                error: Internal server error
                requestId: req_123
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, requestId]
      properties:
        error:
          type: string
        requestId:
          type:
            - string
            - 'null'
...
    OrderStatus:
      type: string
      enum:
        [pending, accepted, cancelled, in_progress, completed_tech, completed]
    OrderFlowType:
      type: string
      enum: [client_selects, tech_applies]
    OperationStatus:
      type: string
      enum: [pending, scheduled, completed_tech, completed, cancelled]
    UsersPageSize:
      type: integer
      enum: [25, 50, 100]
...
    Pagination:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
        pageSize:
          $ref: '#/components/schemas/UsersPageSize'
        total:
          type: integer
        totalPages:
          type: integer
...
    Operation:
      type: object
      required:
        - id
        - order_id
        - client_id
        - technician_id
        - status
        - client_name
        - client_surname
        - client_phone
        - client_whatsapp_phone
        - technician_name
        - technician_surname
        - technician_phone
        - technician_whatsapp_phone
        - order_status
        - order_flow_type
        - service_address_text
        - address_notes
        - zone_slug
        - appliance_type_slug
        - scheduled_at
        - description
        - technician_completed_at
        - completed_at
        - created_at
        - updated_at
        - technician_review
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        technician_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OperationStatus'
        client_name:
          type:
            - string
            - 'null'
        client_surname:
          type:
            - string
            - 'null'
        client_phone:
          type:
            - string
            - 'null'
        client_whatsapp_phone:
          type:
            - string
            - 'null'
        technician_name:
          type:
            - string
            - 'null'
        technician_surname:
          type:
            - string
            - 'null'
        technician_phone:
          type:
            - string
            - 'null'
        technician_whatsapp_phone:
          type:
            - string
            - 'null'
        order_status:
          anyOf:
            - $ref: '#/components/schemas/OrderStatus'
            - type: 'null'
        order_flow_type:
          anyOf:
            - $ref: '#/components/schemas/OrderFlowType'
            - type: 'null'
        service_address_text:
          type:
            - string
            - 'null'
        address_notes:
          type:
            - string
            - 'null'
        zone_slug:
          type:
            - string
            - 'null'
        appliance_type_slug:
          type:
            - string
            - 'null'
        scheduled_at:
          type:
            - string
            - 'null'
          format: date-time
        description:
          type:
            - string
            - 'null'
        technician_completed_at:
          type:
            - string
            - 'null'
          format: date-time
        completed_at:
          type:
            - string
            - 'null'
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        technician_review:
          anyOf:
            - $ref: '#/components/schemas/TechnicianReview'
            - type: 'null'
    TechnicianReview:
      type: object
      required:
        - id
        - rating
        - comment
        - created_at
      properties:
        id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type:
            - string
            - 'null'
        created_at:
          type: string
          format: date-time
    AdminOperation:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          required:
            - order_status
            - order_flow_type
            - service_address_text
            - client_id
            - client_email
            - client_name
            - client_surname
            - technician_email
            - technician_name
            - technician_surname
            - technician_public_slug
          properties:
            order_status:
              $ref: '#/components/schemas/OrderStatus'
            order_flow_type:
              $ref: '#/components/schemas/OrderFlowType'
            service_address_text:
              type: string
            client_id:
              type: string
              format: uuid
            client_email:
              type: string
              format: email
            client_name:
              type:
                - string
                - 'null'
            client_surname:
              type:
                - string
                - 'null'
            technician_email:
              type: string
              format: email
            technician_name:
              type:
                - string
                - 'null'
            technician_surname:
              type:
                - string
                - 'null'
            technician_public_slug:
              type: string
    OperationResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Operation'
    AdminOperationResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/AdminOperation'
    CreateOperationInput:
      type: object
      required:
        - technician_id
        - scheduled_at
      properties:
        technician_id:
          type: string
          format: uuid
        scheduled_at:
          type:
            - string
            - 'null'
          format: date-time
    ScheduleOperationInput:
      type: object
      required:
        - scheduled_at
        - description
      properties:
        scheduled_at:
          type: string
          format: date-time
        description:
          type: string
    PaginatedOperationsSummary:
      type: object
      required:
        [
          totalOperations,
          pendingOperations,
          scheduledOperations,
          completedTechOperations,
          completedOperations,
          cancelledOperations,
        ]
      properties:
        totalOperations:
          type: integer
        pendingOperations:
          type: integer
        scheduledOperations:
          type: integer
        completedTechOperations:
          type: integer
        completedOperations:
          type: integer
        cancelledOperations:
          type: integer
    PaginatedOperations:
      type: object
      required: [items, pagination, summary]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Operation'
        pagination:
          $ref: '#/components/schemas/Pagination'
        summary:
          $ref: '#/components/schemas/PaginatedOperationsSummary'
    PaginatedOperationsResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/PaginatedOperations'
    PaginatedAdminOperations:
      type: object
      required: [items, pagination, summary]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AdminOperation'
        pagination:
          $ref: '#/components/schemas/Pagination'
        summary:
          $ref: '#/components/schemas/PaginatedOperationsSummary'
    PaginatedAdminOperationsResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/PaginatedAdminOperations'
...